add bonus functionality as per Odin tasks
Now the colors are generated randomly when interacting with blocks.
If the interaction with the same block repeats - the opacity reduces by
10%. After 10 clicks the block is fully black and then resets back to
100% opacity.

Changed the gridContainer background to black, so when cell opacity
decreases - it becomes darker.

When colors are reset with the "Reset" button, the opacity also resets to 1.
When the grid is built the opacity value is added and is set to 1.

Fixed the range for Size of the grid. Now its limited to > 0 and < 101.
Removed the brushColor parameter from changeColorOnHover, since its generated
randomly each time the event listener on the cell is triggered.